github-scripts/requirements.txt
backoff==2.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
multidict==6.0.2
python-dateutil==2.8.2
requests==2.28.1
requests-toolbelt==0.10.1
six==1.16.0
urllib3==1.26.12
yarl==1.8.1
